pyproject.toml
[project]
name = "web-to-epub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.5",
    "ebooklib>=0.19",
    "pillow>=11.3.0",
    "python-slugify>=8.0.4",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
]
